Task 4 fixed

+++ sort.h
@@ -33,7 +33,7 @@ int partition(int *array, int first, int last, size_t size);
 void quicksort_recursion(int *array, int first, int last, size_t size);
 void quick_sort(int *array, size_t size);
 void shell_sort(int *array, size_t size);
-void swap_sh(int *arr, int x, int y, int *swapped);
+void swap_sh(int *arr, size_t x, size_t y, size_t *swapped);
 void merge(size_t l, size_t m, size_t r, int *dest, int *org);
 void merge_recrusion(size_t l, size_t r, int *a, int *cpy_arr);
 void merge_sort(int *array, size_t size);
